refactor: break Resume into section components
